Introduce a migration script
The script allows to easilly update the .locres file when a new
update goes live.
TBD: get the list of keys that changed between patches.
Add the Game.locres file which is the up-to-date version of the english
localization (Hotfix #56).
Add base.csv which is a text representation of the .locres.
UnrealLocres.exe allows to import and export .csv files from the .locres
(big thanks to https://github.com/akintos/UnrealLocres)
UE4localizationsTool.exe is used to view and edit .locres files
directly.

Signed-off-by: Serhii Boiko <30057128+sb-lviv@users.noreply.github.com>